@@ -27,7 +27,7 @@
     },
     "minimum-stability": "stable",
     "require": {
-        "php": "^7.1",
+        "php": "^7.4",
         "ext-pcntl": "*",
         "php-ds/php-ds": "^1.2"
     },
@@ -36,6 +36,6 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^7.5",
-        "phpspec/phpspec": "^5.1"
+        "phpspec/phpspec": "^6.0"
     }
 }
